fix: qa types 분리
